aoc/alsa: flush the capture ring buffer before start for voip call
For voip call, both playback and capture ring bufer need to be flushed.
the playback ring buffer is flushed on the AoC side, while the capture
ring buffer has to be flushed on the AP side.

One caveat here is that only one voip start cmd will be executed
regardless which stream (playback/capture) will be started first.

Bug: 197923357

Signed-off-by: Xinhui Zhou <zxinhui@google.com>
Change-Id: Ib1b41478d68867561248ed563744a1670c23c393