Only send channel service power commands to usf service.
The USF service propagates these power commands to the other entities
that need to act on it so there is only a need to send it to the wakeup
USF service and the others will react accordingly.

There is also no need to propagate this message to the USF non-wake-up
channel because USF makes the necessary changes internally when the
command is received in either channel.

Bug: 283488727
Test:  in-device verification.
Change-Id: Ic9dc011c22498acc3cfa79e04397b5352afedc3d